Initial data cleaning is done: HTML removed and unknown characters.
metrics.py contains the initial cleaning code - currently  it reads the file into a pandas dataframe and then parses out the HTML within it.

A side effect seems to be that most of the unknown characters are gone - not sure if all are gone though.

I've added the first initial cleaned data for convenience, but probably won't for later revisions to cut down on size.

A simple batch file is also added to run metrics.py