Achieved 91% accuracy (MORE INFO BELOW)
I was able to get a solid 91% accuracy on the test set (your results may vary due to different versions).
We're still using SGD - I've tried about 8 other classification algorithms and most would give us a much lower accuracy or never actually converge on the data.

One of the biggest things that was changed to achieve the increase of the accuracy was no longer removing stop words. I think we talked about this in one of the meetings that removing the stop words and having bigrams might actually change the meaning of the review and this seems to be the case.

I've begun the construction of a pretty extensive README - after the project is done we can make this repo public.

I've also added an example of sentiment analysis with Keras and TensorFlow. It is based off of an example that Keras has in their documentation and loads in a similar IMDB dataset, but not the same as ours.